Refactors as BetaPage
